pyproject.toml
[tool.poetry]
name = "galileo-callback"
version = "0.1.0"
description = ""
readme = "README.md"
authors = ["Rajma"]
packages = [
    { include = "galileo_callback" }
]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
fastapi = "^0.115.12"
annotated-types = "^0.7.0"
pydantic = "^2.11.4"
uvicorn = "^0.34.2"
httpx = "^0.27.0"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.3"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
